server/requirements.txt
blinker==1.8.2
click==8.1.8
Flask==3.0.3
Flask-Cors==5.0.0
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==2.1.5
psycopg2-binary==2.9.10
Werkzeug==3.0.6
zipp==3.20.2
